mqtt-sparkplug: rough support for data type BYTES
Instead of storing an array of bytes,, data is stored
as an array of objects, whose key is the metric name.

This is so it matches the structure Datasets are stored into.

This is most likely temporary, and will need to be re-worked later,
to keep data as an array of bytes
(storage of Datasets will then be re-worked too)